support non-root paths in base_uri (#62)
* support non-root paths in base_uri

resolves #61

Given a Guzzle client with a base uri like
`http://example.com/myregistryliveshere/`.

Requests where created with a relative path with a leading `/`.
Guzzle interprets this as a path from the root and replaces
`/myregistryliveshere/` with the given path.

To make Guzzle combine the path correctly with the given
base_uri, the relative path must not start with a `/`. Then
it is appended at the end of the base_uri, preserving the
path in the base_uri.

* fixed unit tests for Functions.php

After removing the leading "/" from the requests the unit tests
failed, because they still expected the leading "/". So removed
this from the tests as well to match the code in Functions
again.

Co-authored-by: Mario Eberhard <mario.eberhard@bertschi.com>
Co-authored-by: Thomas Ploch <profiploch@gmail.com>